turn everything into float and begin

+++ src/torrey.h
@@ -24,7 +24,7 @@
 // numerical accuracy as much when we render.
 // Switching to floating point computation is easy --
 // just set Real = float.
-using Real = double;
+using Real = float;
 
 // Lots of PIs!
 const Real c_PI = Real(3.14159265358979323846);
